requirements.txt
Django>=3.2
djangorestframework
django-otp
Pillow
djangorestframework-simplejwt
